+++ doggle-webapp/src/About Page/About.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+function About() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <div>
+          <h2>About page</h2>
+        </div>
+      </header>
+    </div>
+  );
+}
+
+export default About;
