--- doggle-webapp/src/Header.js
@@ -1,46 +1,0 @@
-import React from "react";
-
-function Header() {
-  return (
-    <nav className="navbar navbar-expand-lg navbar-light bg-light">
-      {" "}
-      <div className="container-fluid">
-        {" "}
-        <a className="navbar-brand" href="#">
-          Doggle
-        </a>{" "}
-        <button
-          className="navbar-toggler"
-          type="button"
-          data-bs-toggle="collapse"
-          data-bs-target="#navbarNav"
-          aria-controls="navbarNav"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          {" "}
-          <span className="navbar-toggler-icon"></span>{" "}
-        </button>{" "}
-        <div className="collapse navbar-collapse" id="navbarNav">
-          {" "}
-          <ul className="navbar-nav">
-            {" "}
-            <li className="nav-item">
-              {" "}
-              <a className="nav-link active" aria-current="page" href="#">
-                Home
-              </a>{" "}
-            </li>{" "}
-            <li className="nav-item">
-              {" "}
-              <a className="nav-link" href="#">
-                About
-              </a>{" "}
-            </li>{" "}
-          </ul>{" "}
-        </div>{" "}
-      </div>{" "}
-    </nav>
-  );
-}
-export default Header;
