+++ doggle-webapp/src/Main Pages/ErrorPage.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+function ErrorPage() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <div className="container">
+          <p style={{ fontSize: "200%" }}>ERROR!</p>
+        </div>
+      </header>
+    </div>
+  );
+}
+
+export default ErrorPage;
