+++ doggle-webapp/src/Main Pages/ImageUploadComponent.js
